Extract stdlib and guava type codecs from `:configuration-cache`
To `:stdlib-serialization-codecs` and `:guava-serialization-codecs`
respectively.

Also move more detangled code to `:graph-serialization`.
