+++ .teamcity/subprojects.json
@@ -483,6 +483,13 @@
     "crossVersionTests": false
   },
   {
+    "name": "guava-serialization-codecs",
+    "path": "platforms/core-configuration/guava-serialization-codecs",
+    "unitTests": false,
+    "functionalTests": false,
+    "crossVersionTests": false
+  },
+  {
     "name": "hashing",
     "path": "platforms/core-execution/hashing",
     "unitTests": true,
@@ -1045,6 +1052,13 @@
   {
     "name": "stdlib-kotlin-extensions",
     "path": "platforms/core-configuration/stdlib-kotlin-extensions",
+    "unitTests": false,
+    "functionalTests": false,
+    "crossVersionTests": false
+  },
+  {
+    "name": "stdlib-serialization-codecs",
+    "path": "platforms/core-configuration/stdlib-serialization-codecs",
     "unitTests": false,
     "functionalTests": false,
     "crossVersionTests": false
