+++ platforms/core-configuration/stdlib-serialization-codecs/src/main/kotlin/org/gradle/internal/serialize/codecs/stdlib/ClassCodec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package org.gradle.configurationcache.serialization.codecs
+package org.gradle.internal.serialize.codecs.stdlib
 
 import org.gradle.internal.serialize.graph.Codec
 import org.gradle.internal.serialize.graph.ReadContext
 import org.gradle.internal.serialize.graph.WriteContext
-import java.nio.charset.Charset
 
 
-object CharsetCodec : Codec<Charset> {
+object ClassCodec : Codec<Class<*>> {
 
-    override suspend fun WriteContext.encode(value: Charset) {
-        writeString(value.name())
+    override suspend fun WriteContext.encode(value: Class<*>) {
+        writeClass(value)
     }
 
-    override suspend fun ReadContext.decode(): Charset =
-        Charset.forName(readString())
+    override suspend fun ReadContext.decode(): Class<*>? =
+        readClass()
 }
