+++ platforms/core-configuration/stdlib-serialization-codecs/src/main/kotlin/org/gradle/internal/serialize/codecs/stdlib/CollectionCodecs.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,17 @@
  * limitations under the License.
  */
 
-package org.gradle.configurationcache.serialization.codecs
+package org.gradle.internal.serialize.codecs.stdlib
 
 import org.gradle.internal.serialize.graph.Codec
-import org.gradle.internal.serialize.graph.ReadContext
-import org.gradle.internal.serialize.graph.WriteContext
 import org.gradle.internal.serialize.graph.codec
-import org.gradle.internal.serialize.graph.logPropertyProblem
 import org.gradle.internal.serialize.graph.readCollectionInto
 import org.gradle.internal.serialize.graph.readList
 import org.gradle.internal.serialize.graph.readMapInto
 import org.gradle.internal.serialize.graph.writeCollection
 import org.gradle.internal.serialize.graph.writeMap
-import java.util.Hashtable
 import java.util.ArrayDeque
+import java.util.Hashtable
 import java.util.LinkedList
 import java.util.Properties
 import java.util.TreeMap
@@ -37,19 +34,15 @@ import java.util.concurrent.CopyOnWriteArrayList
 import java.util.concurrent.CopyOnWriteArraySet
 
 
-internal
 val arrayListCodec: Codec<ArrayList<Any?>> = collectionCodec { ArrayList(it) }
 
 
-internal
 val linkedListCodec: Codec<LinkedList<Any?>> = collectionCodec { LinkedList() }
 
 
-internal
 val arrayDequeCodec: Codec<ArrayDeque<Any?>> = collectionCodec { ArrayDeque(it) }
 
 
-internal
 val copyOnWriteArrayListCodec: Codec<CopyOnWriteArrayList<Any?>> = codec(
     { writeCollection(it) },
     {
@@ -60,85 +53,6 @@ val copyOnWriteArrayListCodec: Codec<CopyOnWriteArrayList<Any?>> = codec(
 )
 
 
-/**
- * Decodes HashSet instances as LinkedHashSet to preserve original iteration order.
- */
-internal
-object HashSetCodec : Codec<HashSet<Any?>> {
-
-    override suspend fun WriteContext.encode(value: HashSet<Any?>) {
-        writeCollectionCheckingForCircularElements(value)
-    }
-
-    override suspend fun ReadContext.decode(): HashSet<Any?> =
-        readCollectionCheckingForCircularElementsInto(::LinkedHashSet)
-}
-
-
-private
-suspend fun WriteContext.writeCollectionCheckingForCircularElements(collection: Collection<Any?>) {
-    var writeCircularMarker = true
-    writeCollection(collection) { element ->
-        if (element != null && element in circularReferences && element.javaClass.overridesHashCode()) {
-            logPropertyProblem("serialize") {
-                text("Circular references can lead to undefined behavior upon deserialization.")
-            }
-            if (writeCircularMarker) {
-                write(CircularReferenceMarker.INSTANCE)
-                writeCircularMarker = false
-            }
-        }
-        write(element)
-    }
-}
-
-
-private
-suspend fun <T : MutableCollection<Any?>> ReadContext.readCollectionCheckingForCircularElementsInto(
-    factory: (Int) -> T
-): T {
-    val size = readSmallInt()
-    val container = factory(size)
-    for (i in 0 until size) {
-        val element = read()
-        if (element === CircularReferenceMarker.INSTANCE) {
-            // upon reading a circular reference, wait until all objects have been initialized
-            // before inserting the elements in the resulting set.
-            val remainingSize = size - i
-            val remaining = ArrayList<Any?>(remainingSize).apply {
-                for (j in 0 until remainingSize) {
-                    add(read())
-                }
-            }
-            onFinish {
-                container.addAll(remaining)
-            }
-            return container
-        }
-        container.add(element)
-    }
-    return container
-}
-
-
-private
-object CircularReferenceMarker {
-
-    val INSTANCE: Any = Marker.INSTANCE
-
-    private
-    enum class Marker {
-        INSTANCE
-    }
-}
-
-
-private
-fun Class<*>.overridesHashCode(): Boolean =
-    getMethod("hashCode").declaringClass !== java.lang.Object::class.java
-
-
-internal
 val treeSetCodec: Codec<TreeSet<Any?>> = codec(
     {
         write(it.comparator())
@@ -152,7 +66,6 @@ val treeSetCodec: Codec<TreeSet<Any?>> = codec(
 )
 
 
-internal
 val copyOnWriteArraySetCodec: Codec<CopyOnWriteArraySet<Any?>> = codec(
     { writeCollection(it) },
     {
@@ -163,7 +76,6 @@ val copyOnWriteArraySetCodec: Codec<CopyOnWriteArraySet<Any?>> = codec(
 )
 
 
-internal
 fun <T : MutableCollection<Any?>> collectionCodec(factory: (Int) -> T) = codec(
     { writeCollection(it) },
     { readCollectionInto(factory) }
@@ -173,15 +85,12 @@ fun <T : MutableCollection<Any?>> collectionCodec(factory: (Int) -> T) = codec(
 /**
  * Decodes HashMap instances as LinkedHashMap to preserve original iteration order.
  */
-internal
 val hashMapCodec: Codec<HashMap<Any?, Any?>> = mapCodec { LinkedHashMap(it) }
 
 
-internal
 val linkedHashMapCodec: Codec<LinkedHashMap<Any?, Any?>> = mapCodec { LinkedHashMap(it) }
 
 
-internal
 val concurrentHashMapCodec: Codec<ConcurrentHashMap<Any?, Any?>> = mapCodec { ConcurrentHashMap<Any?, Any?>(it) }
 
 
@@ -189,18 +98,15 @@ val concurrentHashMapCodec: Codec<ConcurrentHashMap<Any?, Any?>> = mapCodec { Co
  * Cannot rely on Java serialization as
  * Hashtable's readObject() calls unsupported ObjectInputStream#readFields().
  */
-internal
 val hashtableCodec: Codec<Hashtable<Any?, Any?>> = mapCodec { Hashtable(it) }
 
 
 /*
  * Decodes Properties as Properties instead of Hashtable.
  */
-internal
 val propertiesCodec: Codec<Properties> = mapCodec { Properties() }
 
 
-internal
 val treeMapCodec: Codec<TreeMap<Any?, Any?>> = codec(
     {
         write(it.comparator())
@@ -214,7 +120,6 @@ val treeMapCodec: Codec<TreeMap<Any?, Any?>> = codec(
 )
 
 
-internal
 fun <T : MutableMap<Any?, Any?>> mapCodec(factory: (Int) -> T): Codec<T> = codec(
     { writeMap(it) },
     { readMapInto(factory) }
