+++ testing/architecture-test/src/test/java/org/gradle/architecture/test/ArchUnitFixture.java
@@ -78,6 +78,7 @@ public interface ArchUnitFixture {
         "org.gradle.internal.configuration.problems..",
         "org.gradle.internal.extensions.core..",
         "org.gradle.internal.flow.services..",
+        "org.gradle.internal.serialize.codecs..",
         "org.gradle.internal.serialize.graph..",
         "org.gradle.kotlin..",
         "org.gradle.internal.declarativedsl..",
